Fix: edit location of CartView
